requirements.txt
python-telegram-bot==21.0
openai>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
